backend/pip_dbus_requirements.txt
asgiref==3.5.2
Bottleneck==1.3.4
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pip==21.2.4
psycopg2==2.8.6
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
scikit-learn==1.1.1
scipy==1.7.3
setuptools==61.2.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==2.2.0
uritemplate==4.1.1
urllib3==1.26.9
wheel==0.37.1
